Rename httpsrv embedded to asset
