+++ back/internal/httpsrv/gogen.go
@@ -1,7 +1,7 @@
 package httpsrv
 
 // alias protobuf/swagger
-//go:generate -command pb-sw protoc -I ../../../msgs/proto -I ../../tools/.include -I ../../../msgs/proto/include/googleapis --openapiv2_out=logtostderr=true,allow_merge=true:internal/embedded/contents
+//go:generate -command pb-sw protoc -I ../../../msgs/proto -I ../../tools/.include -I ../../../msgs/proto/include/googleapis --openapiv2_out=logtostderr=true,allow_merge=true:internal/asset/assets
 
 // generate swagger.json
 //go:generate pb-sw ../../../msgs/proto/auth.proto ../../../msgs/proto/user.proto ../../../msgs/proto/post.proto
