+++ back/internal/httpsrv/internal/asset/asset.go
@@ -1,4 +1,4 @@
-package embedded
+package asset
 
 import (
 	"embed"
@@ -7,12 +7,12 @@ import (
 )
 
 var (
-	//go:embed contents/*
+	//go:embed assets/*
 	FS embed.FS
 )
 
 func NewFS() (fs.FS, error) {
-	s, err := fs.Sub(FS, "contents")
+	s, err := fs.Sub(FS, "assets")
 	if err != nil {
 		return nil, fmt.Errorf("connect to contents fs: %w", err)
 	}
