+++ back/internal/httpsrv/internal/asset/assets/apidocs.swagger.json
